refactor(oracle): extract patch-independent cli ip-fragment helpers into cli/live
